ENH: Make the integer conversions in read and write lengths explicit
Fourth of five commits clearing -Wsign-conversion.  15 hunks on the
length argument of znzread(), znzwrite(), fread() and expat's XML_Parse.

Two of them are worth more than a cast, because the conversion runs the
other way and the declared return type cannot represent the error value:

  znzlib.c   znzread() and znzwrite() return size_t, and both do
      `return nread;` where nread is the int returned by gzread()/
      gzwrite(), documented in a comment as "returns -1 on error".  A
      size_t cannot hold -1: the caller receives SIZE_MAX.  Callers do
      detect it, because they compare the result against the number of
      bytes they asked for, but the comment describes a contract the
      function's type cannot honour.  The conversion is now explicit and
      the comment says what actually reaches the caller.

  nifti1_io.c  nifti_read_buffer() has the same shape, returning -1 from
      a size_t function on two failure paths.

Neither behaviour is changed here -- callers depend on the current
values -- but both are now visible rather than implicit.  Making these
functions report failure in a way their return type can express is a
separate change to their contract.

+++ cifti/afni_xml.c
@@ -282,7 +282,7 @@ axml_read_file(const char * fname, int read_data)
     if (reset_xml_buf(xd, &buf, &bsize))
       break;
 
-    blen = (unsigned)fread(buf, 1, bsize, fp);
+    blen = (unsigned)fread(buf, 1, (size_t)bsize, fp);
 
     /* check for early termination */
     bshort = loc_strnlen(buf, blen);
@@ -298,7 +298,7 @@ axml_read_file(const char * fname, int read_data)
     if (xd->verb > 4)
       fprintf(stderr, "-- XML_Parse # %d\n", pcount);
     pcount++;
-    if (XML_Parse(parser, buf, blen, done) == XML_STATUS_ERROR)
+    if (XML_Parse(parser, buf, (int)blen, done) == XML_STATUS_ERROR)
     {
       fprintf(stderr,
               "** %s at line %u\n",
@@ -394,7 +394,7 @@ axml_read_buf(const char * buf_in, int64_t bin_len)
     if (xd->verb > 4)
       fprintf(stderr, "-- XML_Parse # %d\n", pcount);
     pcount++;
-    if (XML_Parse(parser, buf, blen, done) == XML_STATUS_ERROR)
+    if (XML_Parse(parser, buf, (int)blen, done) == XML_STATUS_ERROR)
     {
       fprintf(stderr,
               "** %s at line %u\n",
