+++ niftilib/nifti1_tool.c
@@ -1933,7 +1933,7 @@ read_file_text(const char * filename, int * length)
     return NULL;
   }
 
-  bytes = fread(text, sizeof(char), len, fp);
+  bytes = fread(text, sizeof(char), (size_t)len, fp);
   fclose(fp); /* in any case */
 
   if (bytes != (size_t)len)
